+++ src/LIAnsureProtect.Worker/Program.cs
@@ -4,8 +4,10 @@ using LIAnsureProtect.Worker;
 
 
 var builder = Host.CreateApplicationBuilder(args);
+var databaseConnectionString = builder.Configuration.GetConnectionString("LIAnsureProtect");
+
 builder.Services.AddApplication();
-builder.Services.AddInfrastructure();
+builder.Services.AddInfrastructure(databaseConnectionString);
 builder.Services.AddHostedService<Worker>();
 
 var host = builder.Build();